+++ src/plugins/core/execute/index.ts
@@ -1,5 +1,4 @@
 export * from './createRequestInterceptor';
-export * from './disabledInterceptor';
 export * from './lazyVariableInterceptor';
 export * from './logResponseInterceptor';
 export * from './processedHttpRegionInterceptor';