+++ src/plugins/core/registerCorePlugin.ts
@@ -28,6 +28,7 @@ function initOnRequestHook(api: models.HttpyacHooksApi) {
   api.hooks.onRequest.addHook('resolveRequestBody', request.resolveRequestBody);
   api.hooks.onRequest.addHook('requestVariableReplacer', request.requestVariableReplacer);
   api.hooks.onRequest.addHook('transformRequestBody', request.transformRequestBodyToBuffer);
+  api.hooks.onRequest.addHook('transfromMultilineFormUrlEncoded', request.transfromMultilineFormUrlEncoded);
   api.hooks.onRequest.addHook('encodeRequestBody', request.encodeRequestBody);
 
   api.hooks.onRequest.addInterceptor(request.isTrustedInterceptor);