+++ src/plugins/core/request/encodeRequestBody.ts
@@ -4,11 +4,7 @@ import * as models from '../../../models';
 import * as utils from '../../../utils';
 
 export async function encodeRequestBody(request: models.Request): Promise<void> {
-  if (request.body) {
-    if (utils.isString(request.body)) {
-      if (utils.isMimeTypeFormUrlEncoded(request.contentType)) {
-        request.body = encodeUrl(request.body);
-      }
-    }
+  if (request.body && utils.isString(request.body) && utils.isMimeTypeFormUrlEncoded(request.contentType)) {
+    request.body = encodeUrl(request.body);
   }
 }