+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fix
 - prevent readFile Error log output on package.json search (Anweber/vscode-httpyac#293)
+- disabled MetaData Handler checks disabled State on occurence of line in execution order (AnWeber/vscode-httpyac#292)
 
 ## [6.13.3]
 ### Fix