+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fix
 - prevent readFile Error log output on package.json search (Anweber/vscode-httpyac#293)
 - disabled MetaData Handler checks disabled State on occurence of line in execution order (AnWeber/vscode-httpyac#292)
+- support Intellij Multiline FormUrlEncoded Syntax (#699)
 
 ## [6.13.3]
 ### Fix